prinout TRs after various thresholds, group thresh changed to >50%, fc w/ dot product fixed

+++ scripts/derive_fc.py
@@ -4,6 +4,7 @@ import sys
 import nibabel as nb
 import hcp_utils as hcp
 import numpy as np
+import os
 
 # function from:
 # https://stackoverflow.com/questions/30143417/computing-the-correlation-coefficient-between-two-multi-dimensional-arrays
@@ -41,36 +42,38 @@ aggregateTS=np.empty([1,59412])
 for T in range(len(tasks)):
 	# load Concat-maskedtime series
 	filepath=parentfp + str(subj) + '_p2mm_masked_filtered_' + tasks[T] + '.dtseries.nii'
-	subjData=nb.load(filepath)
-	# extract cortical TS
-	subjDataCort=subjData.dataobj[:,hcp.struct.cortex]
-	# stack on to aggregate TS
-	aggregateTS=np.append(aggregateTS,subjDataCort,axis=0)
-	# initialize network-level FC matrix
-	NL_fcMat=np.zeros((17,17))
-	# Individual network metrics (FC matrix of all 17 networks, summary-level)
-	for N in range(17):
-		# index of which vertices belong to this network
-		# +1 to resolve python-everything else counting discordance
-		_,NetIndex=np.where(parcelCort==(N+1))
-		NetTS=subjDataCort[:,NetIndex]
-		# index all other networks
-		for OtherNet in range(17):
-			# initialize array of Net * OtherNet corrs (vertexwise)
-			_,OtherNetIndex=np.where(parcelCort==(OtherNet+1))
-			OtherNetTS=subjDataCort[:,OtherNetIndex]
-			# correlations between Net and OtherNet elements
-			Cors=corr2_coeff(NetTS.T,OtherNetTS.T)
-			NL_fcMat[N,OtherNet]=np.mean(Cors)
-		# correct for same-cell autocorrelation: 
-		# subtract ((Vertex#inNet*.5)/NumberOfCells) from mean
-		# .5 because diagonals only show up once in symmetric fc, other elements show up twice
-		WithinCon=NL_fcMat[N,N]
-		WithinConAdjusted=WithinCon-((len(NetIndex)*.5)/((len(NetIndex)*len(NetIndex))))
-		NL_fcMat[N,N]=WithinConAdjusted
-	# save out 17x17 for within-task network stats
-	NLfp=childfp+str(subj)+'_NLFcmat'+tasks[T]
-	np.save(NLfp,NL_fcMat)
+	# if file exists
+	if os.path.exists(filepath):
+		subjData=nb.load(filepath)
+		# extract cortical TS
+		subjDataCort=subjData.dataobj[:,hcp.struct.cortex]
+		# stack on to aggregate TS
+		aggregateTS=np.append(aggregateTS,subjDataCort,axis=0)
+		# initialize network-level FC matrix
+		NL_fcMat=np.zeros((17,17))
+		# Individual network metrics (FC matrix of all 17 networks, summary-level)
+		for N in range(17):
+			# index of which vertices belong to this network
+			# +1 to resolve python-everything else counting discordance
+			_,NetIndex=np.where(parcelCort==(N+1))
+			NetTS=subjDataCort[:,NetIndex]
+			# index all other networks
+			for OtherNet in range(17):
+				# initialize array of Net * OtherNet corrs (vertexwise)
+				_,OtherNetIndex=np.where(parcelCort==(OtherNet+1))
+				OtherNetTS=subjDataCort[:,OtherNetIndex]
+				# correlations between Net and OtherNet elements
+				Cors=corr2_coeff(NetTS.T,OtherNetTS.T)
+				NL_fcMat[N,OtherNet]=np.mean(Cors)
+			# correct for same-cell autocorrelation: 
+			# subtract ((Vertex#inNet*.5)/NumberOfCells) from mean
+			# .5 because diagonals only show up once in symmetric fc, other elements show up twice
+			WithinCon=NL_fcMat[N,N]
+			WithinConAdjusted=WithinCon-((len(NetIndex)*.5)/((len(NetIndex)*len(NetIndex))))
+			NL_fcMat[N,N]=WithinConAdjusted
+		# save out 17x17 for within-task network stats
+		NLfp=childfp+str(subj)+'_NLFcmat'+tasks[T]
+		np.save(NLfp,NL_fcMat)
 
 # remove null column of aggregate TS
 aggregateTS=np.delete(aggregateTS,0,0)
